commit final lab 2

+++ Lab2/src/ro/ase/csie/cts/lab02/interfaces/Profitable.java
@@ -1,6 +1,6 @@
 package ro.ase.csie.cts.lab02.interfaces;
 
 public interface Profitable {
-    public static final float MAX_INTEREST_PERCENT = 5;		// ~ 5%
+    public static final double MAX_INTEREST_PERCENT = 5;
     public abstract void AddInterest(float interestPercent);
 }
